CB1 fixes: MoE aggregation-leader rank + int64 token accessors; fix int32 overflow in prefill attention accumulators
BUGS #29 (route.cpp): GateUpdate::aggregate_expert compared device_total_rank against global rank 0, so under pp>1 any stage whose device_list excludes rank 0 never ran the aggregation (regen + token-conservation assert) at all. Compare against device_list.front() (this stage's own designated aggregator) instead, matching CB1.

BUGS #30 (sequence.h): get_gen_process_token/get_sum_process_token/get_total_sequence_length widened int -> int64_t. They sum per-sequence token counts across an entire batch, which at offline/capacity-search batch sizes can exceed INT32_MAX; matches CB1's int64_t usage.

New: int32 overflow in attention_sum_impl.cpp's flash-path byte accumulators (6 sites: GQA-Sum act_size, and an MLA/Absorb-Sum twin at lines ~944/1121/1122). The (hw_metric) cast wrapped the whole multi-term product/sum expression, so the underlying int arithmetic (m*k*num_heads + m*n*num_heads) could overflow INT32_MAX before ever promoting to double, at large prefill chunk sizes x long context x many heads. Fixed by casting the leading factor of each term instead (matching the file's own already-safe kv_read_size idiom), forcing double arithmetic throughout. Surfaced by an external codebase-integration agent's source audit, independently verified here (confirmed exact line-level mechanism + found 3 additional vulnerable sites beyond the 3 it reported). Confirmed dormant for every paper-1 published number: decode_mode=true means AttentionSum (prefill) is never exercised (BUGS.md #2 -- pushDummySeq forces current_len=input_len at sequence creation, so no sequence ever enters the prefill/sum bucket). Pure correctness hardening.

+++ src/scheduler/sequence.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <chrono>  // for timing of loading sequence
+#include <cstdint>
 #include <ctime>
 #include <fstream>
 #include <iomanip>
@@ -108,10 +109,15 @@ class BatchedSequence {
   void pop(Sequence::Ptr seq);
   int get_num_seq();
   int get_process_token();
-  int get_gen_process_token();
-  int get_sum_process_token();
+  // BUGS #30 (CB1): widened to int64_t -- these sum num_process_token/
+  // current_len (each individually bounded by max_seq_len, safe as int)
+  // across an entire batch, which at the large batch sizes the offline/
+  // capacity-search paths can reach (tens of thousands of sequences x
+  // up to ~131072 current_len) can exceed INT32_MAX and silently wrap.
+  int64_t get_gen_process_token();
+  int64_t get_sum_process_token();
   int get_average_sequence_length();
-  int get_total_sequence_length();
+  int64_t get_total_sequence_length();
 
   void add_dummy_sequence(int num_seq, int input_len, int output_len);
